% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikGenNPH.R
\name{LikGenNPH}
\alias{LikGenNPH}
\title{LikGenNPH}
\usage{
LikGenNPH(y, D, nf, nk, ncl, dist)
}
\arguments{
\item{y}{Vector of parameters in the form
\deqn{y = (ln(a), ln(b), \beta _{shape}, \beta _{scale}, ln(\sigma ^2))}
for Weibull hazard function and
\deqn{y = (ln(10^3a), ln(10^2b), \beta _{shape}, \beta _{scale}, ln(\sigma ^2))}
for Gompertz hazard function, where a and b are slope and shape parameters,
\eqn{\beta _{shape}} and \eqn{\beta _{scale}} are the Cox-regression parameters
 for shape and scale, respectively, and \eqn{\sigma ^2} is the variance of frailty.
 This vector must include at least two parameters, \eqn{ln(a)} and \eqn{ln(b)}.}

\item{D}{A data.frame in which to interpret the variables named in the formula.
The data set includes the following fields:
\enumerate{
\item time-to-failure and censoring in the case without left truncation
or time-of-start, time-of-failure, and censoring in the case with left truncation at the time of begin
(censoring must be either 0 for no event or 1 for event);
\item Covariates (continuous or categorical) used in a study (can be empty set).
}}

\item{nf}{The number of continuous and binary factors in the data set D corresponding to the covariates
used in the Cox-regression for proportional hazard term.}

\item{nk}{The number of continuous and binary factors in the data set D corresponding
to the covariates used in the Cox-regression for shape b.}

\item{ncl}{The number of clusters in the data set D corresponding to the cluster covariate.
Is equal to 0 for the fixed-effect model.}

\item{dist}{Baseline hazard function ('Weibull' or 'Gompertz').}
}
\value{
Neg-loglikelihood
}
\description{
This function calculates the neg-loglikelihood
for the parametric gamma-frailty model with non-proportional hazard functions
}
\examples{
\dontrun{
LikGenNPH(y, D, nf, nk, ncl, dist)
}

}
