% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msurv.R
\name{msurv}
\alias{msurv}
\title{msurv}
\usage{
msurv(D, nf, nk, ncl, dist, vpar, ID)
}
\arguments{
\item{D}{A data.frame in which to interpret the variables named in the formula.
The data set includes the following fields:
\enumerate{
\item time-to-failure and censoring in the case without left truncation
or time-of-start, time-of-failure, and censoring in the case with left truncation at the time of begin
(censoring must be either 0 for no event or 1 for event);
\item Covariates (continuous or categorical) used in a study (can be empty set).
}}

\item{nf}{The number of continuous and binary factors in the data set D corresponding to the covariates
used in the Cox-regression for proportional hazard term.}

\item{nk}{The number of continuous and binary factors in the data set D corresponding
to the covariates used in the Cox-regression for shape \eqn{b}.}

\item{ncl}{The number of clusters in the data set D corresponding to the cluster covariate.
Is equal to 0 for the fixed-effect model).}

\item{dist}{Baseline hazard function ('Weibull' or 'Gompertz').}

\item{vpar}{The sample of the vector-row-parameters in matrix form.}

\item{ID}{The order number of an object in the data set.}
}
\value{
Marginal survival for the object with order number ID
}
\description{
This function calculates marginal survival
}
\examples{
\dontrun{
msurv(D,nf,nk,ncl,dist,vpar,ID)
}

}
