% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamFact.R
\name{NamFact}
\alias{NamFact}
\title{NamFact}
\usage{
NamFact(data, formula.scale, formula.shape)
}
\arguments{
\item{data}{A data.frame in which to interpret the variables named in the formula.
The data set includes the following fields:
\enumerate{
\item Time-to-failure and censoring in the case without left truncation
or time-of-start, time-of-failure, and censoring in the case with left truncation at the time of begin;
Censoring must be either 0 (no event) or 1 (event);
\item Covariates (continuous or categorical) used in a study (can be empty set).
}}

\item{formula.scale}{A formula object, with the response on the left of a ~ operator, and the terms on
the right. The response must be a survival object as returned by the Surv() function.
The status indicator 'event' in the Surv object must be 0=censored, 1=non-censored.
This object describes the effect of several factors in the Cox proportional hazards model.}

\item{formula.shape}{A formula object, with the response on the left of a ~ operator, and the terms on
the right. The response must be a survival object as returned by the Surv() function.
The status indicator 'event' in the Surv object must be 0=censored, 1=non-censored.
This object describes the effect of several factors on the shape parameter.}
}
\value{
List of the factor names
}
\description{
This function returns the list of factor names (after possible converting categorical variables to multiple binary ones as well)
}
\examples{
\dontrun{
NamFact(data, formula.scale, formula.shape)
}

}
